+++ packages/design/src/AgentChatInput.tsx
@@ -6,7 +6,9 @@ import {AgentChatField} from "./agent-chat/Field";
 import {HarnessWidget} from "./agent-chat/HarnessWidget";
 import {AgentChatHint} from "./agent-chat/Hint";
 import {Icon} from "./agent-chat/Icon";
+import {AgentChatOverflow} from "./agent-chat/Overflow";
 import {PiExtensionDialog} from "./agent-chat/PiExtensionDialog";
+import {AgentChatPrimaryActions} from "./agent-chat/PrimaryActions";
 import {type AgentChatInputProps, AgentChatInputRoot, useAgentChatInput} from "./agent-chat/Root";
 import {AgentChatSettings} from "./agent-chat/Settings";
 import {AgentChatSurface} from "./agent-chat/Surface";
@@ -104,3 +106,5 @@ AgentChatInput.Toolbar = AgentChatToolbar;
 AgentChatInput.Settings = AgentChatSettings;
 AgentChatInput.Hint = AgentChatHint;
 AgentChatInput.Control = AgentChatControl;
+AgentChatInput.PrimaryActions = AgentChatPrimaryActions;
+AgentChatInput.Overflow = AgentChatOverflow;
